% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depend_graph.R
\name{add_source_node}
\alias{add_source_node}
\title{Add Source Node To Graph}
\usage{
add_source_node(g)
}
\description{
Add a source node with index 0 for each node without parents, return resulting graph.
}
