% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{apply_location}
\alias{apply_location}
\title{Find Location Of Apply In Parse Tree}
\usage{
apply_location(expr, apply_func = "apply")
}
\arguments{
\item{expr}{expression}

\item{apply_func}{vector of apply names}
}
\value{
integer 1 for \code{apply(x, ...}, 3 for \code{y <- apply(x,
     ...}, 0 otherwise
}
\description{
Only looks at the top level expression.
}
