% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{apply_parallel}
\alias{apply_parallel}
\title{Convert \code{apply} To Parallel}
\usage{
apply_parallel(expr)
}
\arguments{
\item{expression}{\code{apply()} function call}
}
\value{
modified parallel code
}
\description{
Transform a single
\code{apply()} function call into a parallel version using mclapply
}
