% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Assignment-class}
\alias{Assignment-class}
\alias{Assignment}
\title{Assignment}
\description{
Assignment
}
\section{Slots}{

\describe{
\item{\code{lhs}}{name of the variable to be assigned}
}}

