% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{AssignmentOneVectorFunction-class}
\alias{AssignmentOneVectorFunction-class}
\alias{AssignmentOneVectorFunction}
\title{Assignment From Single Vectorized Function}
\description{
Assignment From Single Vectorized Function
}
\section{Slots}{

\describe{
\item{\code{functionName}}{name of the function that's called}

\item{\code{args}}{arguments to the function}
}}

