% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{benchmark_parallel}
\alias{benchmark_parallel}
\title{Run Benchmark To Find Efficient Parallel Code}
\usage{
benchmark_parallel(statement, times = 100L)
}
\arguments{
\item{statement}{a single R statment}

\item{times}{the number of times to run the benchmark}
}
\value{
statement potentially in parallel
}
\description{
One doesn't know ahead of time if it will be more efficient to run code
in serial or in parallel. This function runs the code in both serial and
parallel, then returns the faster version based on the median timing.
}
\details{
It may be nice to return more, or log the profiling
}
