% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSource.R
\name{ChunkDataFiles}
\alias{ChunkDataFiles}
\title{Constructor for ChunkDataFiles}
\usage{
ChunkDataFiles(files, sizes = file.info(files)$size,
  readFuncName = inferReadFuncFromFile(files[1]), ...)
}
\description{
Constructor for ChunkDataFiles
}
