% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{CodeBlock-class}
\alias{CodeBlock-class}
\alias{CodeBlock}
\title{Abstract base class for blocks comprising a DataParallelSchedule}
\description{
These are NOT basic blocks in the sense of compilers, because they may contain control flow.
}
\section{Slots}{

\describe{
\item{\code{code}}{to evaluate in serial on the manager.}
}}

