% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{DataDescription-class}
\alias{DataDescription-class}
\alias{DataDescription}
\title{Description of Data}
\description{
Description of Data
}
\section{Slots}{

\describe{
\item{\code{symbol}}{name of the variable representing the object as used in source code}

\item{\code{fun}}{function that reads or writes the data}

\item{\code{args}}{arguments to fun so that do.call(fun, args) will read or write the data}
}}

