% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{DataFiles-class}
\alias{DataFiles-class}
\alias{DataFiles}
\title{Description of Data Files}
\description{
Contains information necessary to generate a call to read in these data files
}
\section{Slots}{

\describe{
\item{\code{files}}{absolute paths to all the files}

\item{\code{format}}{format of the input files}

\item{\code{Rclass}}{class of the data object in R, for example, \code{"data.frame"}}

\item{\code{details}}{list of details to help efficiently and correctly read in the data}
}}

