% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSource.R
\name{dataFiles}
\alias{dataFiles}
\title{Description of Data Files}
\usage{
dataFiles(dir, files = list.files(dir, full.names = TRUE),
  sizes = sapply(files, function(f) file.info(f)$size),
  readFuncName = "read.csv", ...)
}
\arguments{
\item{dir}{directory filled exclusively with data files}

\item{files}{absolute paths to all the files}

\item{sizes}{sizes of the objects in memory}

\item{...}{further details to help efficiently and correctly read in the data}
}
\value{
\linkS4class{DataFiles}
}
\description{
Contains information necessary to generate a call to read in these data files
}
