% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{DataSource-class}
\alias{DataSource-class}
\alias{DataSource}
\title{Base Class for Data Descriptions}
\description{
Subclasses describe data
}
\section{Slots}{

\describe{
\item{\code{varname}}{character variable name in the original code}
}}

