% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depend_graph.R
\name{depend_graph}
\alias{depend_graph}
\title{Expression Dependency Graph}
\usage{
depend_graph(script, add_source = FALSE)
}
\arguments{
\item{script}{as returned from \code{\link[CodeDepends]{readScript}}}
}
\description{
Create a DAG representing a set of expression dependencies.
I'm not yet trying to take the minimal set of such dependencies, because
it's not clear that this is necessary.
}
