% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{DependGraph-class}
\alias{DependGraph-class}
\alias{DependGraph}
\title{Dependency graph between expressions}
\description{
Dependency graph between expressions
}
\section{Slots}{

\describe{
\item{\code{code}}{input code}

\item{\code{graph}}{data frame representing the graph with indices corresponding
to code}
}}

