% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandData.R
\docType{methods}
\name{expandData,expression,list,ANY-method}
\alias{expandData,expression,list,ANY-method}
\title{Expand Data Description}
\usage{
\S4method{expandData}{expression,list,ANY}(code, data, platform, ...)
}
\description{
Insert the chunked data loading calls directly into the code, expand vectorized function calls,
and collect variables before calling non vectorized function calls.
}
