% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{expandData}
\alias{expandData}
\title{Expand Data}
\usage{
expandData(code, data, platform, ...)
}
\arguments{
\item{code}{file name or a string containing code to be parsed}

\item{data}{list of data descriptions. 
Each element is a \linkS4class{DataSource}.
The names of the list elements correspond to the variables in the code that these objects are bound to.}

\item{platform}{\linkS4class{Platform} describing resource to compute on}

\item{...}{additional arguments to schedule methods}
}
\description{
Updates code to include code to load data
}
