% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ExprChunkData-class}
\alias{ExprChunkData-class}
\alias{ExprChunkData}
\title{Chunked Data Source defined by R expressions}
\description{
Contains information necessary to load chunks of data into an R session.
}
\section{Slots}{

\describe{
\item{\code{expr}}{expression such that evaluating \code{expr[[i]]} produces the ith chunk of data}
}}

