% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/writeCode.R
\docType{methods}
\name{file<-}
\alias{file<-}
\alias{file<-,GeneratedCode,character-method}
\title{Set File for generated code object}
\usage{
file(description) <- value

\S4method{file}{GeneratedCode,character}(description) <- value
}
\arguments{
\item{description}{\linkS4class{GeneratedCode}}

\item{value}{file name to associate with object}
}
\description{
Set File for generated code object
}
