% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{FinalBlock-class}
\alias{FinalBlock-class}
\alias{FinalBlock}
\title{Finalize, shut down the platform, free the resources, because everything is done.}
\description{
Finalize, shut down the platform, free the resources, because everything is done.
}
