% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{find_apply}
\alias{find_apply}
\title{Find Expressions Using \code{apply}}
\usage{
find_apply(expr)
}
\arguments{
\item{expr}{expression}
}
\value{
logical locations of apply
}
\description{
Only looks at the top level expression.
}
