% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSource.R
\name{findFirstDataSource}
\alias{findFirstDataSource}
\title{Find the first expression that we can infer a DataSource from}
\usage{
findFirstDataSource(expr, ...)
}
\description{
Find the first expression that we can infer a DataSource from
}
