% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{FixedWidthFiles-class}
\alias{FixedWidthFiles-class}
\alias{FixedWidthFiles}
\title{A Collection Of One Or More Fixed Width Files}
\description{
A Collection Of One Or More Fixed Width Files
}
