% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ForkSchedule-class}
\alias{ForkSchedule-class}
\alias{ForkSchedule}
\title{Fork based parallel schedule}
\description{
Class for schedules that should be parallelized by forks from one single
process
}
\section{Slots}{

\describe{
\item{\code{sequence}}{vector of statement indices}
}}

