% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{GeneratedCode-class}
\alias{GeneratedCode-class}
\title{Generated code ready to write}
\description{
This class contains code that is ready to run and execute, as well as
the steps taken to generate this code.
}
\section{Slots}{

\describe{
\item{\code{schedule}}{contains all information to generate code}

\item{\code{code}}{executable R code}

\item{\code{file}}{name of a file where code will be written}
}}

