% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{GroupByBlock-class}
\alias{GroupByBlock-class}
\alias{GroupByBlock}
\title{GROUP BY style code to run in parallel}
\description{
GROUP BY style code to run in parallel
}
\section{Slots}{

\describe{
\item{\code{groupData}}{names of chunked variables to split according to groupIndex}

\item{\code{groupIndex}}{names of chunked variables that define the split}
}}

