% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSource.R
\name{inferDataSource}
\alias{inferDataSource}
\title{Infer Data Source Object From Code}
\usage{
inferDataSource(expr)
}
\arguments{
\item{expr}{expression to find the data source in}
}
\description{
Infer Data Source Object From Code
}
