% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{inferPlatform}
\alias{inferPlatform}
\title{Infer Current Platform}
\usage{
inferPlatform()
}
\value{
\linkS4class{Platform}
}
\description{
When generating code in the absence of specific platform information we assume the code will run where it's generated.
}
