% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{InitBlock-class}
\alias{InitBlock-class}
\alias{InitBlock}
\title{Initialize the platform}
\description{
Initialize the platform
}
\section{Slots}{

\describe{
\item{\code{funcNames}}{list of functions defined in slot code to make available for the remainder of the program.}

\item{\code{assignmentIndices}}{assigns each data chunk to a worker. For example, c(2, 1, 1) assigns the 1st chunk to worker 2, and chunks 2 and 3 to worker 1.}
}}

