% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{KnownAssignment-class}
\alias{KnownAssignment-class}
\alias{KnownAssignment}
\title{Simple Statement With Known Value}
\description{
Simple Statement With Known Value
}
\section{Slots}{

\describe{
\item{\code{value}}{the value that the lhs will be bound to}
}}

