% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{MapSchedule-class}
\alias{MapSchedule-class}
\alias{MapSchedule}
\title{Data parallel schedule}
\description{
Class for schedules that should be parallelized with apply style parallelism
}
