% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{MeasuredTaskGraph-class}
\alias{MeasuredTaskGraph-class}
\alias{MeasuredTaskGraph}
\title{Graph where the size of each variable that can be transferred is known}
\description{
Graph where the size of each variable that can be transferred is known
}
