% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depend_graph.R
\name{one_edge}
\alias{one_edge}
\title{Create a single edge from the most recent output less than i, to i}
\usage{
one_edge(i, output)
}
\description{
Create a single edge from the most recent output less than i, to i
}
