% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ParallelBlock-class}
\alias{ParallelBlock-class}
\alias{ParallelBlock}
\title{Code to run in parallel}
\description{
Code to run in parallel
}
\section{Slots}{

\describe{
\item{\code{export}}{names of objects to export from manager to workers.}
}}

