% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelize.R
\name{parallelize}
\alias{parallelize}
\title{Parallelized Data Evaluater}
\usage{
parallelize(varname, cluster = NULL, spec = 2L, ...)
}
\arguments{
\item{varname}{name of an existing list that one expects to use \code{lapply} on}

\item{cluster}{an existing SNOW cluster, or NULL to create one}

\item{spec}{number of workers, see \code{\link[parallel]{makeCluster}}}

\item{...}{additional arguments to \code{\link[parallel]{makeCluster}}}
}
\value{
closure works similarly as \code{eval}
}
\description{
Distributes data over a cluster and returns a closure capable of
evaluating code in parallel. Designed for interactive use.
}
\examples{
x = list(letters, 1:10)
do = parallelize("x")
do(lapply(x, head))
}
