% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{parallelize_script}
\alias{parallelize_script}
\title{Convert A Script To Parallel Through Benchmarking}
\usage{
parallelize_script(expr)
}
\arguments{
\item{expression}{serial code}
}
\value{
modified parallel code
}
\description{
Benchmarking is used to determine if it's worth it go parallel.
}
