% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ParallelLocalCluster-class}
\alias{ParallelLocalCluster-class}
\alias{ParallelLocalCluster}
\title{Placeholder for local \code{cluster} objects from the parallel package, for example, those produced by \code{parallel::makeCluster}.}
\description{
Placeholder for local \code{cluster} objects from the parallel package, for example, those produced by \code{parallel::makeCluster}.
}
\section{Slots}{

\describe{
\item{\code{name}}{symbol to use for the cluster name when generating code}
}}

