% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Platform-class}
\alias{Platform-class}
\alias{Platform}
\title{Description of a Platform}
\description{
Describes the physical and software infrastructure that we can use to generate parallel code.
}
\section{Slots}{

\describe{
\item{\code{workers}}{number of parallel workers to use}
}}

