% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{platform}
\alias{platform}
\title{Describe Platform}
\usage{
platform(OS.type, workers)
}
\arguments{
\item{OS.type}{character, \code{"unix"} or \code{"windows"}}

\item{workers}{integer, number of parallel workers}
}
\value{
\linkS4class{Platform}
}
\description{
Constructor for Platform classes
}
