% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,TaskSchedule,missing-method}
\alias{plot,TaskSchedule,missing-method}
\title{Gantt chart of a schedule}
\usage{
\S4method{plot}{TaskSchedule,missing}(x, blockHeight = 0.25,
  main = "schedule plot", xlab = "Time (seconds)", ylab = "Processor",
  evalColor = "gray", sendColor = "orchid", receiveColor = "slateblue",
  labelTransfer = TRUE, labelExpr = NULL, rectAes = list(density = NA,
  border = "black", lwd = 2), ...)
}
\arguments{
\item{x}{\linkS4class{TaskSchedule}}

\item{blockHeight}{height of rectangle, between 0 and 0.5}

\item{main}{title}

\item{xlab}{x axis label}

\item{ylab}{y ayis label}

\item{evalColor}{color for evaluation blocks}

\item{sendColor}{color for send blocks}

\item{receiveColor}{color for receive blocks}

\item{labelTransfer}{add labels for transfer arrows}

\item{labelExpr}{NULL to use default numbering labels, FALSE to suppress
labels, or a character vector of custom labels.}

\item{rectAes}{list of additional arguments for
\code{\link[graphics]{rect}}}

\item{...}{additional arguments to \code{plot}}
}
\description{
Gantt chart of a schedule
}
