% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ReduceBlock-class}
\alias{ReduceBlock-class}
\alias{ReduceBlock}
\title{Reduce in parallel on the workers}
\description{
Reduce in parallel on the workers
}
\section{Slots}{

\describe{
\item{\code{objectToReduce}}{name of the object to apply the reduce to}

\item{\code{resultName}}{name of the object to save the result as}

\item{\code{reduceFun}}{implementation of a reduce to use}
}}

