% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ReduceFun-class}
\alias{ReduceFun-class}
\alias{ReduceFun}
\title{Abstract base class for reducible function implementations}
\description{
Abstract base class for reducible function implementations
}
\section{Slots}{

\describe{
\item{\code{reduce}}{name of a reducible function}

\item{\code{predicate}}{function that takes in a resource and returns TRUE if this particular resource can be reduced using this ReduceFun, and FALSE otherwise.
TODO: Define resource and make it more user accessible if users are expected to compute on it.}
}}

