% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduceFun}
\alias{reduceFun}
\title{Construct ReduceFun Objects}
\usage{
reduceFun(reduce, summary = reduce, combine = "c", query = summary,
  predicate = function(...) TRUE)
}
\description{
Construct ReduceFun Objects
}
