% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_call}
\alias{replace_call}
\title{Replace Call With Expression}
\usage{
replace_call(input, function_name, replacement)
}
\description{
Replace Call With Expression
}
\examples{
e1 = quote(rnorm(10))
replace_call(e1, "rnorm", 
}
