% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Schedule-class}
\alias{Schedule-class}
\alias{Schedule}
\title{Schedule base class}
\description{
Schedule base class
}
\section{Slots}{

\describe{
\item{\code{graph}}{\linkS4class{DependGraph} used to create the schedule}
}}

