% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/zzz.R
\docType{methods}
\name{schedule}
\alias{schedule}
\alias{schedule,GeneratedCode-method}
\alias{schedule,TaskGraph-method}
\title{Schedule Dependency Graph}
\usage{
schedule(graph, data, platform, ...)

\S4method{schedule}{GeneratedCode}(graph, data, platform, ...)

\S4method{schedule}{TaskGraph}(graph, data, platform = Platform(),
  nWorkers = platform@nWorkers, chunkFuncs = character(),
  reduceFuncs = list(), knownReduceFuncs = getKnownReduceFuncs(),
  knownChunkFuncs = getKnownChunkFuncs(),
  allChunkFuncs = c(knownChunkFuncs, chunkFuncs))
}
\arguments{
\item{graph}{\linkS4class{TaskGraph}, code dependency graph}

\item{data}{list of data descriptions. 
Each element is a \linkS4class{DataSource}.
The names of the list elements correspond to the variables in the code that these objects are bound to.}

\item{platform}{\linkS4class{Platform} describing resource to compute on}

\item{...}{additional arguments to schedule methods}
}
\description{
Creates the schedule for a dependency graph. The schedule is the
assignment of the expressions to different processors at different
times. There are many possible scheduling algorithms. The default is
\code{\link{mapSchedule}}, which does
simple map parallelism using R's apply family of functions.
}
\references{
See \emph{Task Scheduling for Parallel Systems}, Sinnen, O.
for a thorough treatment of what it means to have a valid schedule.
}
