% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduleFork.R
\name{scheduleFork}
\alias{scheduleFork}
\title{Single sequential forks scheduler}
\usage{
scheduleFork(graph, overhead = 1e-03)
}
\arguments{
\item{graph}{\linkS4class{DependGraph} as returned from \code{\link{inferGraph}}}

\item{overhead}{seconds required to initialize a fork}
}
\value{
schedule \linkS4class{ForkSchedule}
}
\description{
Single sequential forks scheduler
}
