% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduleFork.R
\name{scheduleForkSeq}
\alias{scheduleForkSeq}
\title{Single sequential forks scheduler}
\usage{
scheduleForkSeq(graph, overhead)
}
\arguments{
\item{graph}{\linkS4class{DependGraph} as returned from \code{\link{inferGraph}}}

\item{overhead}{seconds required to initialize a fork}
}
\value{
schedule integer representing the schedule. If an index appears
twice, then the first appearance is the fork, and the second is the
join. Indices appearing once mean don't fork.
}
\description{
Single sequential forks scheduler
}
