% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandData.R, R/scheduleTaskList.R
\docType{methods}
\name{expandData,expression,list-method}
\alias{expandData,expression,list-method}
\alias{scheduleTaskList}
\title{Expand Data Description}
\usage{
\S4method{expandData}{expression,list}(code, data, ...)

scheduleTaskList(graph, maxWorker = 2L, orderFun = orderBottomLevel,
  timeDefault = 1e-05,
  sizeDefault = as.numeric(utils::object.size(1L)), overhead = 8e-06,
  bandwidth = 1.5e+09)
}
\arguments{
\item{graph}{\linkS4class{DependGraph} as returned from \code{\link{inferGraph}}}

\item{maxWorker}{integer maximum number of processors}

\item{orderFun}{function that takes in a \code{graph} and
returns a permutation of \code{1:length(graph@code)} that respects the
topological ordering of the graph.}

\item{timeDefault}{numeric time in seconds to execute a single
expression. Expression times default to this value, with a warning, if
we can't find \code{time} from \code{graph}.}

\item{sizeDefault}{numeric default size of objects to transfer in bytes}

\item{overhead}{numeric seconds to send any object}

\item{bandwidth}{numeric speed that the network can transfer an object
between processors in bytes per second. We don't take network
contention into account. This will have to be extended to account for
multiple machines.}
}
\value{
schedule object of class \code{TaskSchedule}
}
\description{
Insert the chunked data loading calls directly into the code, expand vectorized function calls,
and collect variables before calling non vectorized function calls.

Implementation of "list scheduling".
This is a greedy algorithm that assigns each expression to the earliest
possible processor.
}
\details{
This function is experimental and unstable. If you're trying to actually
speed up your code through parallelism then consider using the default
method in \code{\link{schedule}} for data parallelism.
This function rewrites code to use task parallelism.
Task parallelism means two or more processors run different R
expressions simultaneously.
}
\examples{
code <- parse(text = "a <- 100
     b <- 200
     c <- a + b")

g <- inferGraph(code)
s <- scheduleTaskList(g)
plot(s)
}
\references{
Algorithm 10 in \emph{Task Scheduling for Parallel
Systems}, Sinnen (2007)
}
