% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SerialBlock-class}
\alias{SerialBlock-class}
\alias{SerialBlock}
\title{Code to run in serial}
\description{
Code to run in serial
}
\section{Slots}{

\describe{
\item{\code{collect}}{names of objects to collect from the workers to the manager.}
}}

