% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depend_graph.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle vectors x and y together, ie. (x[1], y[1], x[2], y[2], ...)}
\usage{
shuffle(x, y)
}
\description{
Shuffle vectors x and y together, ie. (x[1], y[1], x[2], y[2], ...)
}
