% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SimpleReduceFun-class}
\alias{SimpleReduceFun-class}
\alias{SimpleReduceFun}
\title{Implementation for a reducible function using function names only}
\description{
This assumes that all of the summary, combine and query functions are defined and available in the R environment where it will run.
See \linkS4class{UserDefinedReduceFun} to define and use your own functions.
}
\section{Slots}{

\describe{
\item{\code{summaryFun}}{name of a function that each worker will call on their chunk of the data.
This produces an intermediate result.}

\item{\code{combineFun}}{name of a function to combine many intermediate results into a single intermediate results}

\item{\code{queryFun}}{name of a function to produce the actual final result from an intermediate result}
}}

