% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{simpleReduceFun}
\alias{simpleReduceFun}
\title{Construct ReduceFun Objects}
\usage{
simpleReduceFun(reducibleFunName, summaryFun = reducibleFunName,
  combineFun = "c", queryFun = summaryFun)
}
\description{
Construct ReduceFun Objects
}
