% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SplitBlock-class}
\alias{SplitBlock-class}
\alias{SplitBlock}
\title{Split one chunked object using another as a factor}
\description{
GROUP BY style code becomes a split followed by an lapply, and both are parallel blocks.
The semantic meaning of this in a schedule is that the data will be grouped, ready for an lapply on the groups.
}
\section{Slots}{

\describe{
\item{\code{groupData}}{names of chunked variables to split according to groupIndex}

\item{\code{groupIndex}}{names of chunked variables that define the split}

\item{\code{lhs}}{name of the chunked variable that holds the result of the split.
This doesn't necessarily need to be here, but we use it to generate code.}
}}

