% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardizeData}
\alias{standardizeData}
\title{schedulers expect to see the data in a standard form}
\usage{
standardizeData(data)
}
\description{
schedulers expect to see the data in a standard form
}
