% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Statement-class}
\alias{Statement-class}
\alias{Statement}
\title{Single Top Level Statement}
\description{
Scripts consist of many such statements.
This class is necessary to help out with method dispatch in \code{expandData}.
We would use expression, but there's already a method for that.
}
\section{Slots}{

\describe{
\item{\code{statement}}{language object that is the statement}
}}

