% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{TableChunkData-class}
\alias{TableChunkData-class}
\alias{TableChunkData}
\title{Chunked Tables}
\description{
Chunked Tables
}
\section{Slots}{

\describe{
\item{\code{columns}}{names of the columns}

\item{\code{splitColumn}}{name of the columns by which the data is split.
\code{NA} means no split.}
}}

