% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{TaskGraph-class}
\alias{TaskGraph-class}
\alias{TaskGraph}
\title{Dependency graph between expressions}
\description{
Subclasses of this class contain all the information that we know about
the code and the problem, such as the time to run each expression and
the variable sizes.
}
\section{Slots}{

\describe{
\item{\code{code}}{input code}

\item{\code{graph}}{data frame representing the graph with indices corresponding
to code}
}}

