% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{TaskSchedule-class}
\alias{TaskSchedule-class}
\alias{TaskSchedule}
\title{Task Parallel Schedule}
\description{
Task Parallel Schedule
}
\section{Slots}{

\describe{
\item{\code{transfer}}{transfer variables between processes}

\item{\code{evaluation}}{data.frame assigning expressions to processors}

\item{\code{maxWorker}}{maximum number of processors, similar to \code{mc.cores}
in the parallel package}

\item{\code{overhead}}{minimum time in seconds to evaluate a single expression}

\item{\code{bandwidth}}{network bandwidth in bytes per second}
}}

