% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{TextTableFiles-class}
\alias{TextTableFiles-class}
\alias{TextTableFiles}
\title{Description of Data Files}
\description{
Contains information necessary to generate a call to read in these data files
}
\section{Slots}{

\describe{
\item{\code{files}}{absolute paths to all the files}

\item{\code{varname}}{expected variable name in the code}

\item{\code{details}}{list of details to help efficiently and correctly read in the data}
}}

