% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\docType{methods}
\name{time,TimedDependGraph-method}
\alias{time,TimedDependGraph-method}
\alias{time,Schedule-method}
\title{Expression Run Time}
\usage{
\S4method{time}{TimedDependGraph}(x)

\S4method{time}{Schedule}(x)
}
\arguments{
\item{x}{object containing expression run times}
}
\description{
Extract a numeric vector of expression run times
}
