% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{TimedTaskGraph-class}
\alias{TimedTaskGraph-class}
\alias{TimedTaskGraph}
\title{Graph where the run time for each expression is known}
\description{
Graph where the run time for each expression is known
}
\section{Slots}{

\describe{
\item{\code{time}}{time in seconds to run each expression}
}}

