% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferGraph.R
\name{use_def}
\alias{use_def}
\title{Use Definition Chain}
\usage{
use_def(x, all_uses, all_definitions)
}
\arguments{
\item{x}{variable name}

\item{all_uses}{list containing variable names uses in each expression}

\item{all_definitions}{list containing variable names defined in each expression}
}
\value{
data frame of edges suitable for use with
 \code{\link[igraph]{graph_from_data_frame}}.
}
\description{
Compute a data frame of edges with one edge connecting each use of the
variable x to the most recent definition or update of x.
}
