% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depend_graph.R
\name{vargraph}
\alias{vargraph}
\title{Variable Use Edge Graph}
\usage{
vargraph(varname, used_vars, outputs)
}
\arguments{
\item{varname}{variable name}

\item{used_vars}{list containing variable names used in each expression}

\item{outputs}{list containing variable names defined in each expression}
}
\description{
A vector of edges specifying constraints on evaluation order.
Output is suitable for use with \code{\link[igraph]{make_graph}}.
}
