% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataParallel.R
\name{XXX}
\alias{XXX}
\title{Estimate Time To Execute Function}
\usage{
XXX(maxWorker, sizeInput, sizeOutput)
}
\arguments{
\item{maxWorker}{integer number of parallel workers to use}

\item{sizeInput}{numeric size of each input element in bytes}

\item{sizeOutput}{numeric size of each output element in bytes}
}
\value{
list with the following elements:
\describe{
  \item{serialTime}{Time in seconds to execute the function in serial}
  \item{parallelTime}{Time in seconds to execute the function in
     parallel}
  \item{elementsParallelFaster}{Number of data elements required for a
  parallel version with maxWorker workers to be faster than serial. Can
  be Inf if parallel will never be faster than serial.}
  \item{}{}
}
}
\description{
Estimate Time To Execute Function
}
