% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFiles.R
\name{dataFiles}
\alias{dataFiles}
\title{Description of Data Files}
\usage{
dataFiles(dir, format, Rclass, files = list.files(dir, full.names =
  TRUE), ...)
}
\arguments{
\item{dir}{directory filled exclusively with data files}

\item{format}{format of the input files
TODO: infer this using the files themselves}

\item{Rclass}{class of the data object in R, for example, \code{"data.frame"}}

\item{files}{absolute paths to all the files}

\item{details}{list of details to help efficiently and correctly read in the data}
}
\value{
\linkS4class{DataFiles}
}
\description{
Contains information necessary to generate a call to read in these data files
}
