% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{DataSource-class}
\alias{DataSource-class}
\alias{DataSource}
\title{Abstract Base Class For Data Descriptions}
\description{
Abstract Base Class For Data Descriptions
}
\section{Slots}{

\describe{
\item{\code{varName}}{name of the variable in the code}

\item{\code{uniqueValueBound}}{upper bound for number of distinct values.
TODO: Define what this means for tables and vectors.}
}}

