% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\docType{methods}
\name{dataSource}
\alias{dataSource}
\alias{dataSource,expression,missing-method}
\title{Describe Data Source}
\usage{
dataSource(expr, args, ...)

\S4method{dataSource}{expression,missing}(expr, args, ...)
}
\arguments{
\item{expr}{code to load chunks of the data}
}
\description{
Describe Data Source
}
