% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/dataSource.R
\docType{methods}
\name{dataSource}
\alias{dataSource}
\alias{dataSource,expression,missing-method}
\alias{dataSource,character,vector-method}
\title{Describe Data Source}
\usage{
dataSource(expr, args, varname, ...)

\S4method{dataSource}{expression,missing}(expr, args, varname, ...)

\S4method{dataSource}{character,vector}(expr, args, varname, ...)
}
\arguments{
\item{expr}{code or function to load chunks of the data}

\item{args}{list of arguments to function}

\item{varname}{name of the variable in the source code}

\item{...}{additional arguments to methods}
}
\description{
Creates and returns an object of class DataSource
}
