% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{dataSource}
\alias{dataSource}
\title{Infer Or Extract Data Source Object}
\usage{
dataSource(expr, ...)
}
\arguments{
\item{...}{additional arguments to methods}

\item{x}{object to infer the data source from}
}
\value{
\linkS4class{DataSource} object
}
\description{
Infer Or Extract Data Source Object
}
