% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{expandData}
\alias{expandData}
\title{Expand Data}
\usage{
expandData(code, data, ...)
}
\arguments{
\item{code}{object to update}

\item{data}{description of external data to load}

\item{...}{additional arguments to methods}
}
\description{
Updates code to include code to load data
}
