% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCode.R
\docType{methods}
\name{file,DependGraph-method}
\alias{file,DependGraph-method}
\alias{file,Schedule-method}
\alias{file,GeneratedCode-method}
\title{Get File containing code}
\usage{
\S4method{file}{DependGraph}(description)

\S4method{file}{Schedule}(description)

\S4method{file}{GeneratedCode}(description)
}
\arguments{
\item{description}{object that may have a file associated with it}
}
\description{
Get File containing code
}
