% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/MapSchedule.R, R/generateTask.R
\docType{methods}
\name{generate}
\alias{generate}
\alias{generate,SerialSchedule,ANY,ANY-method}
\alias{generate,MapSchedule,ANY,ANY-method}
\alias{generate,TaskSchedule,ANY,ANY-method}
\title{Generate Code}
\usage{
generate(schedule, platform, data, ...)

\S4method{generate}{SerialSchedule,ANY,ANY}(schedule, platform, data, ...)

\S4method{generate}{MapSchedule,ANY,ANY}(schedule, platform, data, ...)

\S4method{generate}{TaskSchedule,ANY,ANY}(schedule, platform,
  portStart = 33000L, minTimeout = 600)
}
\arguments{
\item{schedule}{object to generate code from, typically an object of class \linkS4class{Schedule}}

\item{platform}{object of class \linkS4class{Platform}}

\item{data}{object of class \linkS4class{DataSource}}

\item{...}{additional arguments to methods}

\item{portStart}{first local port to use, can possibly use up to n * (n -
1) / 2 subsequent ports if every pair of n workers must communicate.}

\item{minTimeout}{timeout for socket connection will be at least this
many seconds.}
}
\value{
x object of class \linkS4class{GeneratedCode}
}
\description{
Produces executable code that relies on a SNOW cluster on a single
machine and sockets.
}
\seealso{
\code{\link{schedule}} generic function to create
\linkS4class{Schedule}, \code{\link{writeCode}} to write and extract the
actual code, and
\code{\link{makeParallel}} to do everything all at once.
}
