% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{inferDataSource}
\alias{inferDataSource}
\title{Infer Data Source Object}
\usage{
inferDataSource(expr, ...)
}
\arguments{
\item{...}{additional arguments to methods}

\item{x}{object to infer the data source from}
}
\value{
\linkS4class{DataSource} object
}
\description{
Infer Data Source Object
}
