% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/inferGraph.R
\docType{methods}
\name{inferGraph}
\alias{inferGraph}
\alias{inferGraph,character,missing-method}
\alias{inferGraph,language,missing-method}
\alias{inferGraph,expression,missing-method}
\alias{inferGraph,ANY,numeric-method}
\title{Infer Task Dependency Graph}
\usage{
inferGraph(code, time, ...)

\S4method{inferGraph}{character,missing}(code, time, ...)

\S4method{inferGraph}{language,missing}(code, time, ...)

\S4method{inferGraph}{expression,missing}(code, time, ...)

\S4method{inferGraph}{ANY,numeric}(code, time, ...)
}
\arguments{
\item{code}{the file path to a script or an object that can be coerced
to an expression.}

\item{time}{time to run each expression}

\item{...}{additional arguments to methods}
}
\value{
object of class \linkS4class{TaskGraph}
}
\description{
Statically analyze code to determine implicit dependencies
}
\examples{
g <- inferGraph(parse(text = "
  a <- 1
  b <- 2
  c <- a + b
  d <- b * c
"))

ig <- as(g, "igraph")
plot(ig)

# To specify the time each expression takes:
g2 <- inferGraph(g@code, time = c(1.1, 2, 0.5, 6))
}
