% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeParallel.R
\name{makeParallel}
\alias{makeParallel}
\title{Create Parallel Code From Serial}
\usage{
makeParallel(code, graph = inferGraph(code), run = FALSE,
  scheduler = schedule, ..., generator = generate, generatorArgs = list(),
  prefix = "gen_", file = TRUE, overWrite = FALSE)
}
\arguments{
\item{code}{file name or expression from \code{\link[base]{parse}}}

\item{graph}{object of class \linkS4class{DependGraph}}

\item{run}{logical, evaluate the code once to gather timings?}

\item{scheduler, }{function to produce a \linkS4class{Schedule}
from a \linkS4class{DependGraph}.}

\item{..., }{additional arguments to scheduler}

\item{generator}{function to produce \linkS4class{GeneratedCode} from a \linkS4class{Schedule}}

\item{generatorArgs}{list of named arguments to use with
\code{generator}}

\item{prefix}{character added to front of file name}

\item{file}{character where to write the generated script. If this is a
logical TRUE and code is a file then use \code{prefix} to make a new
name and write a script if code was a file name. If logical FALSE then
don't write anything to disk.}

\item{overWrite}{logical write over existing generated file}
}
\value{
code object of class \linkS4class{GeneratedCode}
}
\description{
This is the most important function in the package, it performs all the
steps required to generate parallel code. Change the default arguments
to customize how this happens.
By default it writes generated code to a file, pass \code{file = FALSE}
to prevent this.
}
\details{
For more details see the \code{vignette("makeParallel-concepts")}.
}
\examples{
# Try running this on an existing R script to create "gen_script.R"
\dontrun{makeParallel("script.R")}

# All the defaults
d <- makeParallel(parse(text = "lapply(mtcars, mean)"))
writeCode(d)

# Select a different scheduling function
pcode <- makeParallel(parse(text = "x <- 1:100
y <- rep(1, 100)
z <- x + y"), scheduler = scheduleTaskList)

plot(schedule(pcode))
}
