% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{MeasuredDependGraph-class}
\alias{MeasuredDependGraph-class}
\alias{MeasuredDependGraph}
\title{Graph where each expression has been executed, timed, and the size of
the variables have been measured.}
\description{
Will export once full pipeline works.
}
\section{Slots}{

\describe{
\item{\code{time}}{time in seconds to run each expression}
}}

