% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R, R/Platform.R
\name{Platform}
\alias{Platform}
\title{Description of a Platform}
\usage{
Platform(OS.type = .Platform[["OS.type"]],
  nWorkers = parallel::detectCores(), name = "cls", scratchDir = ".")

Platform(OS.type = .Platform[["OS.type"]],
  nWorkers = parallel::detectCores(), name = "cls", scratchDir = ".")
}
\arguments{
\item{OS.type}{character, \code{"unix"} or \code{"windows"}}

\item{nWorkers}{integer, number of parallel workers}
}
\value{
\linkS4class{Platform}
}
\description{
Describes the physical and software infrastructure that we can use to generate parallel code.

Constructor for \linkS4class{Platform} classes, by default uses the current local platform.
}
\section{Slots}{

\describe{
\item{\code{nWorkers}}{number of parallel workers to use}
}}

