% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{platform}
\alias{platform}
\title{extract platform}
\usage{
platform(x, ...)
}
\arguments{
\item{x}{object with a platform}

\item{...}{additional arguments to methods}
}
\value{
\linkS4class{Platform} object
}
\description{
extract platform
}
