% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDOT}
\alias{plotDOT}
\title{Plot Dependency Graph}
\usage{
plotDOT(graph, file = NULL, dotfile = NULL, args = "")
}
\arguments{
\item{graph}{dependGraph}

\item{file}{character where to save pdf image}

\item{dotfile}{character where to save dot commands to produce plot}

\item{args}{character additional arguments to \code{dot} command line program}
}
\description{
Produces a PDF image using graphviz
}
