% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ReduceFun-class}
\alias{ReduceFun-class}
\alias{ReduceFun}
\title{Abstract base class for reducible function implementations}
\description{
Abstract base class for reducible function implementations
}
\section{Slots}{

\describe{
\item{\code{reduce}}{name of a reducible function}
}}

