% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduleFork.R
\name{scheduleFork}
\alias{scheduleFork}
\title{Single sequential forks scheduler}
\usage{
scheduleFork(graph, overhead = 1000, bandwidth = 1.5e+09)
}
\arguments{
\item{graph}{object of class \code{DependGraph} as returned from \code{\link{inferGraph}}
expression.}

\item{overhead}{numeric seconds to send any object}

\item{bandwidth}{numeric speed that the network can transfer an object
between processors in bytes per second. We don't take network
contention into account. This will have to be extended to account for
multiple machines.}
}
\value{
schedule object of class \code{ForkSchedule}
}
\description{
TODO: Inherit parameters from scheduleTaskList
}
