% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{TableChunkData-class}
\alias{TableChunkData-class}
\alias{TableChunkData}
\title{Chunked Data Source defined by R expressions}
\description{
Contains information necessary to load chunks of data into an R session.
}
\section{Slots}{

\describe{
\item{\code{expr}}{expression such that evaluating \code{expr[[i]]} produces the ith chunk of data.
Requires evaluating parent expressions first.}

\item{\code{columns}}{names of the columns}

\item{\code{splitColumn}}{name of the columns by which the data is split.
\code{NA} means no split.}

\item{\code{mangledNames}}{names of each chunk of data}

\item{\code{collector}}{name of a function to call to collect all the chunks into one object}

\item{\code{collected}}{for internal use with \code{expandData}}
}}

